+++ lib/features/exercise/data/datasources/i_exercise_datasource.dart
@@ -4,5 +4,6 @@ import '../models/exercise_model.dart';
 
 abstract class IExerciseDatasource {
   Future<Unit> add(ExerciseModel exercise);
+  Future<Unit> remove(List<int> ids);
   Future<List<ExerciseModel>> get exercises;
 }