+++ lib/features/exercise/domain/repositories/i_exercise_repository.dart
@@ -6,4 +6,5 @@ import '../entities/exercise_failure.dart';
 abstract class IExerciseRepository {
   Future<Either<ExerciseFailure, Unit>> add(Exercise exercise);
   Future<Either<ExerciseFailure, List<Exercise>>> get();
+  Future<Either<ExerciseFailure, Unit>> remove(List<int> ids);
 }