+++ lib/features/exercise/presentation/bloc/exercise_event.dart
@@ -21,3 +21,22 @@ class ExerciseFetchEvent extends ExerciseEvent {}
 class OnboardingDoneEvent extends ExerciseEvent {}
 
 class OnboardingIsDoneEvent extends ExerciseEvent {}
+
+class ExerciseSelectionModeEvent extends ExerciseEvent {
+  final bool isInSelectionMode;
+  final List<int> selectedIds;
+
+  const ExerciseSelectionModeEvent({@required this.isInSelectionMode, @required this.selectedIds});
+
+  @override
+  List<Object> get props => [isInSelectionMode, selectedIds];
+}
+
+class ExerciseRemoveEvent extends ExerciseEvent {
+  final List<int> ids;
+
+  const ExerciseRemoveEvent({@required this.ids});
+
+  @override
+  List<Object> get props => [ids];
+}