+++ lib/features/exercise/presentation/pages/dashboard/dashboard_page.dart
@@ -8,37 +8,43 @@ import '../../../domain/entities/exercise.dart';
 import '../../bloc/exercise_bloc.dart';
 import '../../widgets/centered_loading.dart';
 import '../../widgets/pp_appbar.dart';
+import 'widgets/dummy_card.dart';
+import 'widgets/exercise_card.dart';
 
 class DashboardPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: BlocListener<ExerciseBloc, ExerciseState>(
+      body: BlocConsumer<ExerciseBloc, ExerciseState>(
+        buildWhen: (previous, current) {
+          return current is! ExerciseSelectionModeState;
+        },
         listener: (context, state) {
           if (state is ExerciseErrorState) {
             Scaffold.of(context).showSnackBar(SnackBar(content: Text(state.message)));
           }
         },
-        child: BlocBuilder<ExerciseBloc, ExerciseState>(
-          builder: (context, state) {
-            // fetch exercises on initial state or when an exercise gets added
-            if (state is! ExerciseFetchedState || state is ExerciseAddedState) {
-              WidgetsBinding.instance.addPostFrameCallback((_) {
-                BlocProvider.of<ExerciseBloc>(context).add(ExerciseFetchEvent());
-              });
-            }
+        builder: (context, state) {
+          // fetch exercises on initial state or when an exercise gets added
+          if (state is! ExerciseFetchedState || state is ExerciseAddedState) {
+            WidgetsBinding.instance.addPostFrameCallback((_) {
+              BlocProvider.of<ExerciseBloc>(context).add(ExerciseFetchEvent());
+            });
+          }
 
-            // show exercises when they're loaded
-            if (state is ExerciseFetchedState) {
-              return _Body(exercises: state.exercises);
-            }
+          // show exercises when they're loaded
+          if (state is ExerciseFetchedState) {
+            return _Body(exercises: state.exercises);
+          }
 
-            return CenteredLoading();
-          },
-        ),
+          return CenteredLoading();
+        },
       ),
       appBar: PPAppBar(
         titleLabel: 'Dashboard',
+        actions: [
+          _RemoveButton(),
+        ],
       ),
       floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
       floatingActionButton: FloatingActionButton(
@@ -48,15 +54,14 @@ class DashboardPage extends StatelessWidget {
         child: const Icon(Icons.add),
       ),
       bottomNavigationBar: BottomAppBar(
+        elevation: 10,
         shape: const CircularNotchedRectangle(),
-        notchMargin: 4.0,
+        notchMargin: 10.0,
         child: Row(
-          mainAxisSize: MainAxisSize.max,
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: <Widget>[
-            IconButton(
-              icon: const Icon(Icons.menu),
-              onPressed: () {},
+            Container(
+              height: 40,
             ),
           ],
         ),
@@ -65,122 +70,90 @@ class DashboardPage extends StatelessWidget {
   }
 }
 
-class _Body extends StatelessWidget {
+class _Body extends StatefulWidget {
   final List<Exercise> exercises;
 
   const _Body({Key key, this.exercises}) : super(key: key);
 
   @override
+  _BodyState createState() => _BodyState();
+}
+
+class _BodyState extends State<_Body> {
+  final List<int> selectedExerciseIds = [];
+
+  bool isInSelectionMode = false;
+
+  void select(int id, BuildContext context) {
+    setState(() {
+      if (selectedExerciseIds.contains(id)) {
+        selectedExerciseIds.removeWhere((element) => element == id);
+      } else {
+        selectedExerciseIds.add(id);
+      }
+
+      isInSelectionMode = selectedExerciseIds.isNotEmpty;
+
+      context.bloc<ExerciseBloc>().add(ExerciseSelectionModeEvent(
+          isInSelectionMode: isInSelectionMode, selectedIds: selectedExerciseIds));
+    });
+  }
+
+  bool isSelected(int id) => selectedExerciseIds.contains(id);
+
+  @override
   Widget build(BuildContext context) {
     return Padding(
       padding: const EdgeInsets.all(16.0),
-      child: ListView.builder(
-        scrollDirection: Axis.vertical,
-        shrinkWrap: true,
-        itemCount: exercises.length,
-        itemBuilder: (context, index) => _ExerciseCard(exercise: exercises[index]),
-      ),
-    );
-  }
-}
-
-class _DummyCard extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      color: Colors.grey.shade100,
-      elevation: 0,
-      child: InkWell(
-        onTap: () {
-          Navigator.of(context).pushNamed(RoutePaths.exerciseAdd);
-        },
-        child: Container(
-          height: 100,
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: const [
-              Icon(
-                Icons.add,
-                color: Colors.grey,
-              ),
-              Text(
-                'Add new',
-                style: TextStyle(color: Colors.grey),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _ExerciseCard extends StatelessWidget {
-  final Exercise exercise;
-
-  const _ExerciseCard({Key key, @required this.exercise}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      color: Colors.white,
-      elevation: 1,
-      child: InkWell(
-        onTap: () {},
-        child: Container(
-          height: 100,
-          child: Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      child: widget.exercises.isEmpty
+          ? ListView(
+              scrollDirection: Axis.vertical,
+              shrinkWrap: true,
               children: [
-                Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  mainAxisAlignment: MainAxisAlignment.spaceAround,
-                  children: [
-                    Text(
-                      exercise.name.getOrCrash(),
-                      style: Theme.of(context).textTheme.headline6.apply(
-                            color: Colors.grey.shade700,
-                          ),
-                    ),
-                    Text(
-                      '${exercise.oneRm.getOrCrash()} Kg',
-                      style: Theme.of(context).textTheme.subtitle1.apply(
-                            color: Colors.black54,
-                          ),
-                    ),
-                    Row(
-                      crossAxisAlignment: CrossAxisAlignment.center,
-                      mainAxisAlignment: MainAxisAlignment.start,
-                      children: [
-                        Icon(
-                          Icons.keyboard_arrow_right,
-                          color: Theme.of(context).accentColor,
-                        ),
-                        Text(
-                          'Accumulation',
-                          style: Theme.of(context).textTheme.bodyText1.apply(
-                                color: Theme.of(context).accentColor,
-                              ),
-                        ),
-                      ],
-                    )
-                  ],
-                ),
-                Column(children: [
-                  Text(
-                    'Month 1',
-                    style: Theme.of(context).textTheme.subtitle2.apply(
-                          color: Colors.black54,
-                        ),
-                  ),
-                ])
+                DummyCard(),
               ],
+            )
+          : ListView.builder(
+              scrollDirection: Axis.vertical,
+              shrinkWrap: true,
+              itemCount: widget.exercises.length,
+              itemBuilder: (context, index) => ExerciseCard(
+                key: Key(index.toString()),
+                onSelect: () {
+                  select(widget.exercises[index].id, context);
+                },
+                exercise: widget.exercises[index],
+                isInSelectionMode: isInSelectionMode,
+                isSelected: isSelected(widget.exercises[index].id),
+              ),
             ),
-          ),
-        ),
-      ),
+    );
+  }
+}
+
+class _RemoveButton extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return BlocBuilder<ExerciseBloc, ExerciseState>(
+      condition: (previous, current) {
+        if (current is ExerciseFetchedState) return false;
+        return true;
+      },
+      builder: (context, state) {
+        if (state is ExerciseSelectionModeState) {
+          if (state.isInSelectionMode) {
+            return IconButton(
+              icon: const Icon(Icons.delete),
+              color: Colors.black,
+              onPressed: () {
+                context.bloc<ExerciseBloc>().add(ExerciseRemoveEvent(ids: state.selectedIds));
+              },
+            );
+          }
+        }
+
+        return Container();
+      },
     );
   }
 }