+++ CHANGELOG.md
@@ -2,6 +2,48 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.26.0](https://github.com/KiritchoukC/power_progress/compare/v1.25.0...v1.26.0) (2020-05-13)
+
+
+### Features
+
+* **exercises-remove:** add repository and datasource implementation ([08b87f8](https://github.com/KiritchoukC/power_progress/commit/08b87f8)), closes [#6](https://github.com/KiritchoukC/power_progress/issues/6)
+
+## [1.25.0](https://github.com/KiritchoukC/power_progress/compare/v1.24.0...v1.25.0) (2020-05-13)
+
+
+### Features
+
+* **exercise-remove:** implement UI, BLoC and Usecase for exercise removing ([ffa4ac7](https://github.com/KiritchoukC/power_progress/commit/ffa4ac7)), closes [#6](https://github.com/KiritchoukC/power_progress/issues/6)
+
+## [1.24.0](https://github.com/KiritchoukC/power_progress/compare/v1.23.1...v1.24.0) (2020-05-13)
+
+
+### Features
+
+* **card-selection:** implement card selection feature on long press ([7a50a2c](https://github.com/KiritchoukC/power_progress/commit/7a50a2c)), closes [#5](https://github.com/KiritchoukC/power_progress/issues/5)
+
+### [1.23.1](https://github.com/KiritchoukC/power_progress/compare/v1.23.0...v1.23.1) (2020-05-09)
+
+
+### Bug Fixes
+
+* **appbar:** remove useless icon ([40118ea](https://github.com/KiritchoukC/power_progress/commit/40118ea))
+
+## [1.23.0](https://github.com/KiritchoukC/power_progress/compare/v1.22.0...v1.23.0) (2020-05-09)
+
+
+### Features
+
+* **onboarding-skip:** add a skip button to the onboarding welcome page ([7cf89f4](https://github.com/KiritchoukC/power_progress/commit/7cf89f4))
+
+## [1.22.0](https://github.com/KiritchoukC/power_progress/compare/v1.21.0...v1.22.0) (2020-05-09)
+
+
+### Features
+
+* **dummy-card:** add a dummy card when there is no exercise yet ([dae8a26](https://github.com/KiritchoukC/power_progress/commit/dae8a26))
+
 ## [1.21.0](https://github.com/KiritchoukC/power_progress/compare/v1.20.0...v1.21.0) (2020-05-09)